[DEV] Post: CRUD operations
- added flash messages

+++ src/main/java/nikitakirin/forum/entity/Post.java
@@ -15,6 +15,6 @@ public class Post {
 
     @Column(columnDefinition = "TEXT")
     private String description;
-    private int creationDatetime;
+    private Long creationDatetime;
 
 }
